Add local download upload (#3)
Add basic download and upload functionality.
For the download, calls pymix /sync endpoint and downloads the diff between what's on the server and what's in the client's: "/Users/lukepurnell/Library/Application Support/subbox/music" directory. The idea here is to make it as simple as possible for users to download their updated music with the least amount of network traffic by syncing the diff between their local files that electron has access to and what's on the subbox server.

For the upload, uploads single files using filebrowser api.

Several TODOs remain:

- the download feature currently syncs with the drive, this requires use of the FS and therefore electron. Could also add an option to download the user's entire library as a browser download which could be done straight from the web app.
- only show sync feature if running in electron (otherwise grey out and lead user to download the desktop app)
- support configuring some of the hardcoded paths (i.e. the /Application Support file path where downloads are stored to)
- add support for downloading the rekordbox and serato info